add seperate component for priority cell
