+++ src/components/Columns.tsx
@@ -1,15 +1,7 @@
 import { ColumnDef } from '@tanstack/react-table'
 import { TodosData } from 'models/todos'
-import { MoreHorizontal } from 'lucide-react'
 
-import { Button } from '@/components/ui/button'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
-import useUpdateTodo from '@/hooks/use-update-todo'
+import PriorityCell from './PriorityCell'
 
 export const columns: ColumnDef<TodosData>[] = [
   {
@@ -23,41 +15,7 @@ export const columns: ColumnDef<TodosData>[] = [
   {
     accessorKey: 'priority',
     header: 'Priority',
-    cell: ({ row }) => {
-      const todo = row.original
-      const updateTodo = useUpdateTodo()
-
-      const priorityClickHandler = (newPriority: string) => {
-        const updatedTodo = updateTodo.mutate({
-          ...todo,
-          priority: newPriority,
-        })
-        console.log('updatedTodo:', updatedTodo)
-      }
-
-      return (
-        <div className="flex justify-between items-center">
-          <div>{todo.priority}</div>
-          <div>
-            <DropdownMenu>
-              <DropdownMenuTrigger>
-                <Button variant="ghost" className="h-8 w-8 p-0">
-                  <span className="sr-only">Open menu</span>
-                  <MoreHorizontal className="h-4 w-4" />
-                </Button>
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="end">
-                <DropdownMenuItem onClick={() => priorityClickHandler('low')}>
-                  low
-                </DropdownMenuItem>
-                <DropdownMenuItem>medium</DropdownMenuItem>
-                <DropdownMenuItem>high</DropdownMenuItem>
-              </DropdownMenuContent>
-            </DropdownMenu>
-          </div>
-        </div>
-      )
-    },
+    cell: PriorityCell,
   },
   {
     accessorKey: 'dueDate',
